+++ src/sensors/fps_v2_range_5v.c
@@ -2,7 +2,7 @@
 #include "include/sensors.h"
 #include "include/blast_data.h"
 
-int fps_v2_range_5v(double *ptr, mraa_context context)
+int fps_v2_range_5v(float *ptr, mraa_context context)
 {
 	mraa_aio_context aio = (mraa_aio_context)context;
